+++ lib/preprocess_image.py
@@ -1,9 +1,9 @@
 import numpy
 
+INPUT_SIZE = (64, 64)
 
-async def pimg(image):
-  image = image.resize((64, 64))
-  image = numpy.array(image)
-  image = image / 255.0
-  image = numpy.expand_dims(image, axis=0)
-  return image
+
+def pimg(image):
+  image = image.resize(INPUT_SIZE)
+  image = numpy.array(image, dtype=numpy.float32) / 255.0
+  return numpy.expand_dims(image, axis=0)